✨feat: Bar 컴포넌트로 기존 기능 이동, 라벨 추가
