+++ src/components/FinalResultChart/Bar.tsx
@@ -1,0 +1,82 @@
+import {
+  Tooltip,
+  TooltipContent,
+  TooltipProvider,
+  TooltipTrigger,
+} from '@/components/Tooltip';
+
+interface BarProps {
+  candidate1: string;
+  votes1: number;
+  candidate2: string;
+  votes2: number;
+  nonParticipants: number;
+}
+
+const Bar = ({
+  votes1,
+  candidate1,
+  votes2,
+  candidate2,
+  nonParticipants,
+}: BarProps) => {
+  const totalVotes = votes1 + votes2;
+
+  const percentageCandidate1 = (votes1 / totalVotes) * 100;
+  const percentageCandidate2 = (votes2 / totalVotes) * 100;
+  const percentageNonParticipants = (nonParticipants / totalVotes) * 100;
+
+  return (
+    <div className="w-full flex items-center justify-center gap-6">
+      <span>{candidate1}</span>
+
+      <div className="flex h-8 min-w-[50%]">
+        <TooltipProvider>
+          <Tooltip>
+            <TooltipTrigger
+              className="bg-primary rounded-l-full flex items-center justify-center"
+              style={{ width: `${percentageCandidate1}%` }}
+            >
+              <span className="font-bold bg-clip-text text-transparent bg-gradient-purple text-center w-full">{`${votes1}`}</span>
+            </TooltipTrigger>
+            <TooltipContent>
+              <p>{percentageCandidate1.toFixed(1)}%</p>
+            </TooltipContent>
+          </Tooltip>
+        </TooltipProvider>
+
+        <TooltipProvider>
+          <Tooltip>
+            <TooltipTrigger
+              className="bg-tertiary flex items-center justify-center" // 미참여자 색상 변경
+              style={{ width: `${percentageNonParticipants}%` }}
+            >
+              <span className="font-bold bg-clip-text text-transparent bg-gradient-purple text-center w-full">{`${nonParticipants}`}</span>
+            </TooltipTrigger>
+            <TooltipContent>
+              <p>{percentageNonParticipants.toFixed(1)}%</p>
+            </TooltipContent>
+          </Tooltip>
+        </TooltipProvider>
+
+        <TooltipProvider>
+          <Tooltip>
+            <TooltipTrigger
+              className="bg-secondary rounded-r-full flex items-center justify-center"
+              style={{ width: `${percentageCandidate2}%` }}
+            >
+              <span className="font-bold bg-clip-text text-transparent bg-gradient-purple text-center w-full">{`${votes2}`}</span>
+            </TooltipTrigger>
+            <TooltipContent>
+              <p>{percentageCandidate2.toFixed(1)}%</p>
+            </TooltipContent>
+          </Tooltip>
+        </TooltipProvider>
+      </div>
+
+      <span>{candidate2}</span>
+    </div>
+  );
+};
+
+export default Bar;
