+++ src/components/FinalResultChart/FinalResultChart.tsx
@@ -1,72 +1,22 @@
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipProvider,
-  TooltipTrigger,
-} from '@/components/Tooltip';
+import Bar from './Bar';
 
-interface FinalResultChartProps {
-  votes1: number;
-  votes2: number;
-  nonParticipants: number;
-}
-
-const FinalResultChart = ({
-  votes1,
-  votes2,
-  nonParticipants,
-}: FinalResultChartProps) => {
-  const totalVotes = votes1 + votes2;
-
-  const percentageCandidate1 = (votes1 / totalVotes) * 100;
-  const percentageCandidate2 = (votes2 / totalVotes) * 100;
-  const percentageNonParticipants = (nonParticipants / totalVotes) * 100;
-
+const FinalResultChart = () => {
   return (
-    <div className="w-full max-w-xs mx-auto">
-      <div className="flex h-8">
-        <TooltipProvider>
-          <Tooltip>
-            <TooltipTrigger
-              className="bg-primary rounded-l-full flex items-center justify-center"
-              style={{ width: `${percentageCandidate1}%` }}
-            >
-              <span className="font-bold bg-clip-text text-transparent bg-gradient-purple text-center w-full">{`${votes1}`}</span>
-            </TooltipTrigger>
-            <TooltipContent>
-              <p>{percentageCandidate1.toFixed(1)}%</p>
-            </TooltipContent>
-          </Tooltip>
-        </TooltipProvider>
-
-        <TooltipProvider>
-          <Tooltip>
-            <TooltipTrigger
-              className="bg-tertiary flex items-center justify-center" // 미참여자 색상 변경
-              style={{ width: `${percentageNonParticipants}%` }}
-            >
-              <span className="font-bold bg-clip-text text-transparent bg-gradient-purple text-center w-full">{`${nonParticipants}`}</span>
-            </TooltipTrigger>
-            <TooltipContent>
-              <p>{percentageNonParticipants.toFixed(1)}%</p>
-            </TooltipContent>
-          </Tooltip>
-        </TooltipProvider>
-
-        <TooltipProvider>
-          <Tooltip>
-            <TooltipTrigger
-              className="bg-secondary rounded-r-full flex items-center justify-center"
-              style={{ width: `${percentageCandidate2}%` }}
-            >
-              <span className="font-bold bg-clip-text text-transparent bg-gradient-purple text-center w-full">{`${votes2}`}</span>
-            </TooltipTrigger>
-            <TooltipContent>
-              <p>{percentageCandidate2.toFixed(1)}%</p>
-            </TooltipContent>
-          </Tooltip>
-        </TooltipProvider>
-      </div>
+    <div className="bg-container-600">
+      <Bar
+        candidate1="강아지"
+        votes1={2}
+        candidate2="고양이"
+        votes2={7}
+        nonParticipants={1}
+      ></Bar>
+      <Bar
+        candidate1="강하띠"
+        votes1={4}
+        candidate2="코앵이"
+        votes2={6}
+        nonParticipants={0}
+      ></Bar>
     </div>
   );
 };
